add missing editor configuration, linting, req packages
To ensure that there is consistency in style in all of our related codebases, nc-multiplex should also use the same linting
rules as the core `netcreate-2018` and `netcreate-itest` modules. Currently these files are most up-to-date in itest,
as 2018 had linting completely disabled which helped break code formatting consistency. The new itest rules are much more
forgiving now and deliver good results, so they have been added to this repo to fix the many inconsistencies that are in it.